splitting-embedding/requirements.txt
llama-index
llama-index-core
llama-index-embeddings-huggingface
llama-index-embeddings-vertex
llama-index-llms-groq
llama-index-llms-vertex
llama-index-vector-stores-vertexaivectorsearch
google-cloud-aiplatform
python-dotenv

